requirements.txt
opencv-python==4.10.0.84
mediapipe==0.10.14
numpy==1.26.4
scikit-learn==1.5.2
matplotlib==3.9.2
Pillow==10.4.0
googletrans==4.0.0-rc1
gtts==2.5.3
playsound==1.3.0
